refactor: move stepped form components to shared UI package
- Move SteppedFormHeader and SteppedFormNavigationButtons from vbpay and vbcall to @workspace/ui
- Update all imports across both apps to use shared components
- Remove duplicate local component copies
- Consolidate SteppedFormStep type definition in shared package
- Maintain exact same functionality while eliminating code duplication
