+++ apps/vbcall/app/(authed)/admin/plans/components/health-plan-form/health-plan-form-step-values.ts
@@ -1,6 +1,6 @@
 import { File, IdCard } from "lucide-react";
 
-import { SteppedFormStep } from "@/types/stepped-form-step";
+import { SteppedFormStep } from "@workspace/ui/types/stepped-form-step";
 
 export const HealthPlanFormStepValues: SteppedFormStep[] = [
   { id: 1, title: "Basic Information", icon: File },
