+++ apps/vbpay/app/(authed)/admin/settings/components/admin-settings-form-steps.ts
@@ -1,6 +1,6 @@
 import { IdCard, Settings, SquareFunction } from "lucide-react";
 
-import { SteppedFormStep } from "@/types/stepped-form-step";
+import { SteppedFormStep } from "@workspace/ui/types/stepped-form-step";
 
 export const AdminSettingsFormSteps: SteppedFormStep[] = [
   { id: 1, title: "Global Settings", icon: Settings },
