+++ apps/vbpay/app/(authed)/admin/settings/components/stepped-admin-settings-form.tsx
@@ -3,13 +3,13 @@
 import { ErrorDialog } from "@workspace/ui/components/error-dialog";
 import { Form } from "@workspace/ui/components/form";
 import { SheetHeader, SheetTitle } from "@workspace/ui/components/sheet";
+import { SteppedFormHeader } from "@workspace/ui/components/stepped-form-header";
+import { SteppedFormNavigationButtons } from "@workspace/ui/components/stepped-form-navigation-buttons";
 
 import { SetupFormData } from "@/components/setup-form/setup-form-schema";
 import { SetupStep1LicenseInfo } from "@/components/setup-form/steps/setup-step1-license-info";
 import { SetupStep2LicenseFunctionality } from "@/components/setup-form/steps/setup-step2-license-functionality";
 import { SetupStep3GlobalSettings } from "@/components/setup-form/steps/setup-step3-global-settings";
-import { SteppedFormHeader } from "@/components/stepped-form-header";
-import { SteppedFormNavigationButtons } from "@/components/stepped-form-navigation-buttons";
 
 import { useSteppedAdminSettingsForm } from "../hooks/use-stepped-admin-settings-form";
 import {
