+++ apps/vbpay/app/(authed)/network/payer/[slug]/components/py-config/payer-py-config-form-steps.ts
@@ -1,6 +1,6 @@
 import { ClipboardList, File } from "lucide-react";
 
-import { SteppedFormStep } from "@/types/stepped-form-step";
+import { SteppedFormStep } from "@workspace/ui/types/stepped-form-step";
 
 export const PayerPyConfigFormSteps: SteppedFormStep[] = [
   { id: 1, title: "Basic Information", icon: File },
