+++ apps/vbpay/app/(authed)/setup/components/setup-form-steps.ts
@@ -1,6 +1,6 @@
 import { IdCard, Settings, SquareFunction } from "lucide-react";
 
-import { SteppedFormStep } from "@/types/stepped-form-step";
+import { SteppedFormStep } from "@workspace/ui/types/stepped-form-step";
 
 export const SetupFormSteps: SteppedFormStep[] = [
   { id: 1, title: "License Information", icon: IdCard },
