--- apps/vbpay/types/stepped-form-step.ts
@@ -1,5 +1,0 @@
-export type SteppedFormStep = {
-  id: number;
-  title: string;
-  icon: React.ElementType;
-};
